requirements.txt
# requirements.txt
streamlit==1.28.0
pandas==1.5.3
plotly==5.17.0
PyMuPDF==1.23.5
python-docx==0.8.11
spacy==3.6.1
scikit-learn==1.3.0
nltk==3.8.1
textstat==0.7.3